+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Update dependencies:
   - [#540](https://github.com/influxdata/influxdb-client-csharp/pull/540): `Microsoft.Extensions.ObjectPool` to `7.0.8`
 
 #### Examples:
-  - [#542](https://github.com/influxdata/influxdb-client-csharp/pull/542): `Radzen.Blazor` to `4.13.3`
+  - [#551](https://github.com/influxdata/influxdb-client-csharp/pull/551): `Radzen.Blazor` to `4.13.6`
 
 #### Test:
   - [#536](https://github.com/influxdata/influxdb-client-csharp/pull/536): `Microsoft.NET.Test.Sdk` to `17.6.2`